Directory live capture of all text files
Improve directory live capture for monitoring server logs.
- Enable directory live capture for non-json text files.
- Limit the monitoring to files with TXT and LOG extensions, so old log
files with date at the end will be skipped.
- Global default setting to include all text files.

For more flexibility, we might want to consider a per tab option to
toggle include all text files or not.  Let's see if it's worth the extra
UI and code complexity after using this for awhile.