Copy items to clipboard in Excel friendly format (#23)
Change the Ctrl/Cmd-C clipboard copying of items to tab-delimited single
line format. And add a header row as well. Makes it easier to paste in
table format.

Limit the filter text in status bar to ~100 chars, so it doesn't cover
the whole bar and hide other messages.

Refactor StatusBar class to simplify construction and set right padding margin itself.