Cache highlight colors for faster and smoother interactions (#25)
* Copy items to clipboard in Excel friendly format

Change the Ctrl/Cmd-C clipboard copying of items to tab-delimited single
line format. And add a header row as well. Makes it easier to paste in
table format.

Limit the filter text in status bar to ~100 chars, so it doesn't cover
the whole bar and hide other messages.

* Set margin in status bar right label

Per Luis suggestion, change UI to set correct padding margin in status
bar, so we don't need to set them manually in messages.

* Cache highlight colors for faster and smoother interactions

This addresses a UI perf issue encountered: using a large highlight
filter set (12 filters with 5 regex) on a single large event (75K chars)
is enough to cause bad UI freezes. It's addressed with:
- Cache the highlight filter colors of each top level tree item.
- Skip the filter check for non top level tree items.
- Refactor the TreeModel: add proper accessors of the highlight filters
so cache will be reset as filters changed.

* Remove commented code