Better layout for scaling and fix file checks on Windows (#26)
* Fix for bug: same file can be opened in multiple tabs on Windows.
Bug: the same file or directory can be opened in separate tabs on Windows, when the paths are specified in different casing.
Fix: Always use lower casing on Windows to maintain list of opened paths and to compare.

* UI tweaks for better layout in 125-150% scaling on Windows
The main label on Save Filters dialog was cut off when displayed with 125-150% scaling on Windows. Simplify the title and labels and enlarge some sizes to address this.
For scaling more than 150%, it would require setting dpiawareness=0 at command line or conf file.