Copy items as a HTML table
Change the copy to clipboard function to include HTML table format, in
addition to the tab separated text format.
This makes it easier to copy and paste items for bug reporting.
Most apps, including Outlook, Excel and Chrome, have options to paste it
as HTML (default) and as text.
But for more versatility, also add context menu to allow copy as text
only.
Both types of copy can be pasted to Tableau directly.

Binaries available at internal installation point.