Add gray background theme option.
In lieu of a dark theme, add an option to show log view and value dialog
in gray background to reduce contrast.
For simplicity, the UI option is a single check box; but for future
flexibility, the option is saved as a color string instead of a Boolean
value in the settings file.
If desired, we can allow setting different color in the UI later. The
code also recognized if the color is “hacked” in the settings file and
avoid overwriting it.