Query info viz (#43)
* Remember list of last monitored log folders.

Keep track of folders opened for live capture to make it easier to
monitor non-desktop log folders, like UT logs or server logs:
1. Include folder paths in the most recent file list; increase the max
to 15 items; enable live directory capture from there.
2. Also save the last open folder to the INI file, so “Choose directory”
always opens the same folder by default across instances.

* CPP file

* Add gray background theme option.

In lieu of a dark theme, add an option to show log view and value dialog
in gray background to reduce contrast.
For simplicity, the UI option is a single check box; but for future
flexibility, the option is saved as a color string instead of a Boolean
value in the settings file.
If desired, we can allow setting different color in the UI later. The
code also recognized if the color is “hacked” in the settings file and
avoid overwriting it.

* One more file

* Query info viz

New feature to parse log files for query information, output them to CSV
file and provide a Tableau workbook to view and analyze them.

* Added clear all events button to menu bar (#41)

* Update the version of Qt used in Linux from 5.7 to 5.7.1 (#42)

* Adddress linux gcc error with init non-const ref with rvalue

* Fix data source directory in TWB

Remove user specfic directory used in the workbook.
Fix a regex with lazy matching.