Add Evergreen State theme (#46)
* Add Evergreen State theme

Add a mid-darkness and mid-contrast green theme.
Refactor Theme class a bit to decouple it with the rest of project, so
adding a theme just requires a free function and an entry to the map, no
need to edit or recompile other files.

* Handle unsupported theme

If the setting/ini file has an unsupported theme, instead of AV on
launch, we will now revert quitely to Native theme. Same for options
dialog.  The unsupported theme will remain in the setting file, as it
might still be supported in other version of TLV.