Show time without date if whole log spans only a single day
So far, timestamps were always displayed including the date.
This commit adds a new option to hide the date and to only display the
time of a log event. By default, we now display only the time if the all
messages in the log happened on the same day. If the log spans multiple
days, we also include the date.
With this change, we reduce visual clutter and make it easier to focus
on the relevant bits of information.