Support for multiple selected events in "Highligh all events of this type"
So far, "Highligh all events of this type" was only available when only
a single event was selected. This commit adds support for highlighting
the types of multiple selected events.

Each type which is currently selected will get highlighted with its own
color. Also, there will be no duplicate highlight filters added for types
which are already highlighted.