"Highlight all events of this thread/session/request/...
This commit adds support for highlighting the current session,
request and severity. As with "Highlight all events of this type",
the user can select multiple sessions/requests/severities to highlight
at the same time.

Highlighting the current site, user or process might be valuable, too.
But I did not yet run into a scenario where I would have needed it.