Add support for multiple selected events to `Hide this event` (#62)
`Hide this event` and `Hide all events of this type` now support
multiple selected events. Not that I really use this functionality.
The purpose of this commit is mostly to make the behavior consistent
with `Highlight all events of this type` for which support for
multiple selected events was added recently.