Refactoring: make ownership over TreeModel and TreeView more explicit
The treeview and the model owned by the LogTab were accessed from
different places by traversing the QT widget tree. At first, this was
not obvious to me. This commit makes it more clear which other classes
access the contents of a logtab by introducing explicit getters on the
LogTab