Fix multi item hide/delete (#73)
Multi deletion had a very strange behavior. Sometimes it would remove some items (not necessarily the one you selected) and sometimes it would just crash TLV.

This is because every time we remove a row, the indices are updates. So if we remove index 5 and 6, we might actually be deleting index 5 and 7.

At first I thought about just doing the loop in reverse since I though the selected indices are always sorted. However, doing selection with Ctrl+click is also valid and, in this case, order is not guaranteed.

Seems like the easiest solution is to sort the selected indices in reverse order and then remove the items (higher index should be removed first).

Some discussion in the Qt forum: http://www.qtcentre.org/threads/32721-How-can-I-remove-a-list-of-selected-items-in-the-QListView-in-QT-4-6