Fix to automatically display the datetime format when log spans multiple days (#80)
he code in LogTab::InitTreeView that determines whether to show the datetime OR time format by default was not working correctly.

It was assuming to be getting DateTime from the treemodel (from the COL::Time column). However, it was really getting an integer with the milliseconds since the epoch.

I fixed the TreeModel; the Qt::UserRole no longer has a special code path for the Time column. The "ShowRowDeltas" code had to be updated accordingly.