Allow choosing the notation directly from the value dialog
Inspired by Luis' recent addition of YAML and JSON formating for events,
this commit adds a dropdown to the value dialog (the dialog which shows
the event contents) which allows to choose the desired format.

While the global setting was pretty useful, I soon realized that I had
to switch between the different display formats too often, as different
use cases require different formats:
* if I want to execute a Hyper query plan, I first have to change to
  the JSON view
* looking at a logical plan required me to switch either to the YAML or
  flat format
* copy/pasting a SQL query also requires the YAML/flat format

With this change, the used format can be changed directly within the
value dialog. The user's choice is properly persisted, so that the
next time the dialog is opened, the same notation is selected.
The global setting is no longer used for the value dialog but it is
still in use for formatting the event list.