Update appveyor and some source files to Qt 5.11 and MSVC 2017 (#94)
Apparently Qt 5.7 reached EOL and is removed from AppVeyor. Time to jump to latest 5.11. But 5.11 requires MSVC 2017 to use QtWebEngine, and also two include file changes.
Note: Changing the build environment to MSVC 2017 in the AppVeyor project does not force it to use a new build worker. So it needs to be set in the YML file.