Time-only for live views; Update NormalizeLogText; Syntax highlight for msg. (#108)
* Default live view to show time only.

Default to show time only instead of date and time, when there are  less than two initial events, such as in live directory view.

* Update NormalizeLogText.

Update for new IDs and use vector for ordered list.

* Allow syntax highlight for msg

Remove syntax highlight restriction on k=msg. It just depends on if the value is structured json value, like a msg with error code.